requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
amocrm-api==2.6.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
black==23.7.0
build==0.10.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
Django==4.2.3
docker==6.1.3
environs==9.5.0
flake8==6.1.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
jsonfield==3.1.0
magic-filter==1.0.9
marshmallow==3.19.0
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pip-tools==7.3.0
platformdirs==3.8.1
psycopg2-binary==2.9.6
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
websocket-client==1.6.2
yarl==1.9.2
